+++ pojo/src/main/java/com/ybk/dto/match/EventDTO.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
  */
 @Data
 public class EventDTO {
-    private Long eventId;
+    private Integer eventId;
     // 活动名称
     private String name;
     // 活动描述
